+++ ippid.c
@@ -28,28 +28,28 @@
 /*==========================================================================*/
 
 /* Local variables. */
-static double lastError;    /**< Store position error.                      */
-static double iTerm;        /**< Store integral term.                       */
-static double error;        /**< Error between two measurement.             */
-static double pTerm;        /**< Proportionnal error.                       */
-static double dTerm;        /**< Derivate error.                            */
-static double pidValue;     /**< PID value, sum of all the errors.          */
-static double pidLeft;      /**< PID result for Left motor.                 */
-static double pidRight;     /**< PID result for Right motor.                */
+static float lastError;    /**< Store position error.                       */
+static float iTerm;        /**< Store integral term.                        */
+static float error;        /**< Error between two measurement.              */
+static float pTerm;        /**< Proportionnal error.                        */
+static float dTerm;        /**< Derivate error.                             */
+static float pidValue;     /**< PID value, sum of all the errors.           */
+static float pidLeft;      /**< PID result for Left motor.                  */
+static float pidRight;     /**< PID result for Right motor.                 */
 
-static double kp = 55.468;  /**< Proportional parameter of PID corrector.   */
-static double ki = 0.554;   /**< Integral parameter of PID corrector.       */
-static double kd = 42.524;  /**< Derivate parameter of PID corrector.       */
+static float kp = 55.468;  /**< Proportional parameter of PID corrector.    */
+static float ki = 0.554;   /**< Integral parameter of PID corrector.        */
+static float kd = 42.524;  /**< Derivate parameter of PID corrector.        */
 
-static double velocityScaleStop     = 30; /**< Max velocity of the robot.   */
-static double velocityScaleTurning  = 35; /**< Max turning velocity.        */
+static float velocityScaleStop     = 30; /**< Max velocity of the robot.    */
+static float velocityScaleTurning  = 35; /**< Max turning velocity.         */
 
 static int16_t zoneA = 4000;          /**< Area to ajust robot PID.         */
 static int16_t zoneB = 2000;          /**< Area to ajust robot PID.         */
-static double positionScaleA = 250;   /**< One resolution is 464 pulses.    */
-static double positionScaleB = 500;   /**< Max position scale for control.  */
-static double positionScaleC = 1000;  /**< Max position scale for control.  */
-static double velocityScaleMove = 35; /**< Velocity scale use to move.      */
+static float positionScaleA = 250;   /**< One resolution is 464 pulses.     */
+static float positionScaleB = 500;   /**< Max position scale for control.   */
+static float positionScaleC = 1000;  /**< Max position scale for control.   */
+static float velocityScaleMove = 35; /**< Velocity scale use to move.       */
 
 bool steerForward;          /**< Robot oriantation forward.                 */
 bool steerBackward;         /**< Robot oriantation backward.                */
@@ -79,19 +79,19 @@ extern long targetPosition;
  * @param[in] offset     angle we want to add to the target angle
  * @param[in] turning    value use to turn robot over rigth or left
 */
-void pid(double pitch, double restAngle, double offset, double turning) {
+void pid(float pitch, float restAngle, float offset, float turning) {
 
   /* Steer robot forward. */
   if (steerForward) {
     /* Scale down offset at high speed and scale up when reversing */
-    offset    += (double)wheelVelocity/velocityScaleMove;
+    offset    += (float)wheelVelocity/velocityScaleMove;
     restAngle -= offset;
   }
 
   /* Steer robot backward. */
   else if (steerBackward) {
     /* Scale down offset at high speed and scale up when reversing */
-    offset    -= (double)wheelVelocity/velocityScaleMove;
+    offset    -= (float)wheelVelocity/velocityScaleMove;
     restAngle += offset;
   }
 
@@ -99,12 +99,12 @@ void pid(double pitch, double restAngle, double offset, double turning) {
   else if (steerStop) {
     long positionError = wheelPosition - targetPosition;
     if (abs(positionError) > zoneA)                       /* Inside zone A. */
-      restAngle -= (double)positionError/positionScaleA;
+      restAngle -= (float)positionError/positionScaleA;
     else if (abs(positionError) > zoneB)                  /* Inside zone B. */
-      restAngle -= (double)positionError/positionScaleB;
+      restAngle -= (float)positionError/positionScaleB;
     else                                                  /* Inside zone C. */
-      restAngle -= (double)positionError/positionScaleC;
-    restAngle -= (double)wheelVelocity/velocityScaleStop;
+      restAngle -= (float)positionError/positionScaleC;
+    restAngle -= (float)wheelVelocity/velocityScaleStop;
     if (restAngle < 160) // Limit rest Angle
       restAngle = 160;
     else if (restAngle > 200)
@@ -122,7 +122,7 @@ void pid(double pitch, double restAngle, double offset, double turning) {
   /* Steer robot sideways. */
   if (steerLeft) {
     /* Scale down at high speed */
-    turning -= abs((double)wheelVelocity/velocityScaleTurning);
+    turning -= abs((float)wheelVelocity/velocityScaleTurning);
     if (turning < 0)
       turning = 0;
     pidLeft   = pidValue-turning;
@@ -130,7 +130,7 @@ void pid(double pitch, double restAngle, double offset, double turning) {
   }
   else if (steerRight) {
     /* Scale down at high speed */
-    turning -= abs((double)wheelVelocity/velocityScaleTurning);
+    turning -= abs((float)wheelVelocity/velocityScaleTurning);
     if (turning < 0)
       turning = 0;
     pidLeft   = pidValue+turning;
